Added Stopwords trait
This trait will include a new method `stopwords`. This is a simple wrap
on top of the `stopwords` crate. Each library has their own list of
languages, this trait makes the usage of both libraries easy.

If a given language does not have the list of stopwords an empty HashSet
is returned instead.